src/dashboard_django/requirements_django.txt
# Django Dashboard Requirements
# Installer avec: pip install -r requirements_django.txt

Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
whitenoise>=6.6
pandas>=2.0
numpy>=1.24
gunicorn>=21.0
